+++ example/example.dart
@@ -16,8 +16,7 @@ import 'package:sponge_client_dart/sponge_client_dart.dart';
 
 void main() async {
   // Create a new client for an anonymous user.
-  var client = SpongeRestClient(
-      SpongeRestClientConfiguration('http://localhost:8888'));
+  var client = SpongeClient(SpongeClientConfiguration('http://localhost:8888'));
 
   // Get the Sponge server version.
   var version = await client.getVersion();
@@ -37,8 +36,8 @@ void main() async {
   print('Sent event id: $eventId.');
 
   // Create a new client for a named user.
-  client = SpongeRestClient(
-    SpongeRestClientConfiguration('http://localhost:8888')
+  client = SpongeClient(
+    SpongeClientConfiguration('http://localhost:8888')
       ..username = 'john'
       ..password = 'password',
   );